smart_maps_scraper/requirements.txt
streamlit
selenium
pandas
requests
python-dotenv
